Update License
TLDR; you can use it to play the game, create addons, monetizes streams or other content created through playing the game with the mod or modify the project for personal use. you cannot claim the mod as your own or monetize access to the project.

I have also added "abandonement terms" which (shortened view license for full terms) basically means that if i abandon the project than anyone can do whatever they want with it. This is mostly in place for the event of my death (i hope thats a long ways away) or otherwise departure of the minecraft modding scene.